Coverity: Fix CID1393976
Cast to u64 to prevent possible integer overflow.

Reviewed-on: https://review.coreboot.org/27447
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/cavium/cn81xx/clock.c
@@ -63,7 +63,7 @@ u64 thunderx_get_io_clock(void)
 
 	rst_boot.u = read64((void *)RST_PF_BAR0);
 
-	return rst_boot.s.pnr_mul * PLL_REF_CLK;
+	return ((u64)rst_boot.s.pnr_mul) * PLL_REF_CLK;
 }
 
 /**
@@ -75,5 +75,5 @@ u64 thunderx_get_core_clock(void)
 
 	rst_boot.u = read64((void *)RST_PF_BAR0);
 
-	return rst_boot.s.c_mul * PLL_REF_CLK;
+	return ((u64)rst_boot.s.c_mul) * PLL_REF_CLK;
 }
